started rebuilding the algorithm
